Corrected small inconsistencies (no functional changes)
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@639 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ui/aui_common/aui_sound.h
@@ -53,7 +53,7 @@ public:
     );
 
 protected:
-	void TrimWavHeader( uint8 **wavdata, uint32 *size );
+	void TrimWavHeader( uint8 **wavdata, size_t *size );
 };
